rmm: Add support for virtio
This commit enables realm to use virtio (e.g., file system access) by
allowing shared pages to be accessed by the Host. With this commit,
realm-linux will be able to boot to the shell with islet.

[Memory sharing between Realm and the Host]

Realm can modify its private region into a shared region by invoking
`RSI_IPA_STATE_SET` with RIPAS:empty (ref. Figure D2.1 Realm shared
memory protocol flow in rmm-spec). With the following `RMI_RTT_SET_RIPAS`,
this triggers a set of actions by the Host (e.g., `RTT_SET_RIPAS`->
`RTT_READ_ENTRY`->`DATA_DESTROY`) according to the value of RTT entry
for freeing regions (For more about the detail, please refer to
realm_tear_down_rtt_range() in nw-linux). This commit fills this flow.

[Additional Notes]

Currently, `islet`'s page table traversing (e.g., ipa_to_pa()) in default
checks its validity with bit 0. However, this cannot be done in stage 2
page table, as bit [0-1] is used as description type and valid page table
entry has that bit on. In this commit, `ipa_to_pte()` does not conduct
the bit 0 based validity check (here `no_valid_check`: true is passed in
`entry()`).

Also, page table entry modification in default attaches unintended flags which
is not compatible with stage 2 table entry (e.g., RIPAS, HIPAS field).
To avoid incorrect bits to be set, I've added `is_raw` parameter in PTE
modification APIs, which when set as true, only sets the flags that
the users have provided.

+++ rmm/monitor/src/rmi/rec/handlers.rs
@@ -91,7 +91,7 @@ pub fn set_event_handler(mainloop: &mut Mainloop) {
         rmi.receive_gic_state_from_host(rec.rd.id(), rec.id(), &run)?;
         rmi.emulate_mmio(rec.rd.id(), rec.id(), &run)?;
 
-        let ripas = rec.ripas_addr();
+        let ripas = rec.ripas_addr() as usize;
         if ripas > 0 {
             rmi.set_reg(rec.rd.id(), rec.id(), 0, 0)?;
             rmi.set_reg(rec.rd.id(), rec.id(), 1, ripas)?;
